@@ -3,7 +3,7 @@ description: Riverpod's AsyncValue.
 publish_to: 'none'
 
 environment:
-  sdk: '>=2.17.0 <3.0.0'
+  sdk: '>=3.5.0 <4.0.0'
 
 dependencies:
   riverpod: ^2.0.0
