@@ -2,8 +2,8 @@ name: command_store
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
-  flutter: ">=3.0.0"
+  sdk: '>=3.5.0 <4.0.0'
+  flutter: '>=3.24.0'
 
 dependencies:
   accel_key:
@@ -18,4 +18,4 @@ dev_dependencies:
   flutter_lints: ^2.0.0
   flutter_test:
     sdk: flutter
-  mockito: 5.4.0
+  mockito: '>=5.4.0 <5.5.0'
