@@ -3,8 +3,8 @@ description: Context menu.
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
-  flutter: ">=3.0.0"
+  sdk: '>=3.5.0 <4.0.0'
+  flutter: ">=3.24.0 <4.0.0"
 
 dependencies:
   flutter:
