@@ -2,7 +2,7 @@ name: jsettings_example
 publish_to: 'none'
 
 environment:
-  sdk: '>=2.18.0 <3.0.0'
+  sdk: '>=3.5.0 <4.0.0'
 
 dependencies:
   collection: ^1.16.0
@@ -13,8 +13,8 @@ dependencies:
   path: ^1.8.2
   provider: ^6.0.4
   ubuntu_service: ^0.2.0
-  xdg_directories: ^0.2.0+2
-  yaru: ^0.9.0
+  xdg_directories: ^1.0.0
+  yaru: ^6.0.0
 
 dev_dependencies:
   flutter_lints: ^2.0.0
