@@ -6,8 +6,8 @@ issue_tracker: https://github.com/ubuntu-flutter-community/jsettings.dart/issues
 version: 0.1.0
 
 environment:
-  sdk: '>=2.17.0 <3.0.0'
-  flutter: '>=3.0.0'
+  sdk: '>=3.5.0 <4.0.0'
+  flutter: '>=3.24.0'
 
 dependencies:
   flutter:
@@ -19,4 +19,4 @@ dev_dependencies:
   flutter_lints: ^2.0.0
   flutter_test:
     sdk: flutter
-  mockito: ^5.3.2
+  mockito: '>=5.4.0 <5.5.0'
