@@ -6,7 +6,7 @@ issue_tracker: https://github.com/jpnurmi/jsettings.dart/issues
 version: 0.1.1
 
 environment:
-  sdk: '>=2.17.0 <3.0.0'
+  sdk: '>=3.5.0 <4.0.0'
 
 dependencies:
   collection: ^1.16.0
@@ -14,5 +14,5 @@ dependencies:
 
 dev_dependencies:
   lints: ^2.0.0
-  mockito: ^5.3.2
+  mockito: ">=5.4.0 <5.5.0"
   test: ^1.16.0
