@@ -7,21 +7,21 @@ description: |
 homepage: https://github.com/canonical/lxd.dart
 
 environment:
-  sdk: '>=2.17.0 <3.0.0'
+  sdk: '>=3.5.0 <4.0.0'
 
 platforms:
   linux:
 
 dependencies:
   collection: ^1.15.0
-  freezed_annotation: ^2.0.3
-  json_annotation: ^4.5.0
+  freezed_annotation: ^2.4.4
+  json_annotation: ^4.9.0
   meta: ^1.7.0
 
 dev_dependencies:
   build_runner: ^2.1.10
-  freezed: ^2.3.1
-  json_serializable: ^6.2.0
+  freezed: ^2.5.7
+  json_serializable: ^6.8.0
   lints: ^2.0.0
-  mockito: 5.4.0
+  mockito: ">=5.4.0 <5.5.0"
   test: ^1.16.8
