@@ -2,7 +2,7 @@ name: lxd_service
 publish_to: 'none'
 
 environment:
-  sdk: '>=2.17.0 <3.0.0'
+  sdk: '>=3.5.0 <4.0.0'
 
 dependencies:
   collection: ^1.16.0
@@ -24,5 +24,5 @@ dev_dependencies:
   freezed: ^2.3.1
   json_serializable: ^6.2.0
   lints: ^2.0.0
-  mockito: 5.4.0
+  mockito: ">=5.4.0 <5.5.0"
   test: ^1.16.0
