@@ -2,7 +2,7 @@ name: lxd_test
 publish_to: 'none'
 
 environment:
-  sdk: '>=2.17.0 <3.0.0'
+  sdk: '>=3.5.0 <4.0.0'
 
 dependencies:
   lxd:
