@@ -5,8 +5,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:provider/provider.dart';
-import 'package:yaru_icons/yaru_icons.dart';
-import 'package:yaru_widgets/yaru_widgets.dart';
+import 'package:yaru/icons.dart';
+import 'package:yaru/widgets.dart';
 
 import 'instance_store.dart';
 import 'instance_table_model.dart';
@@ -81,8 +81,8 @@ class _InstanceTableViewState extends State<_InstanceTableView> {
           autofocus: true,
           child: Scaffold(
             body: DataTable2(
-              headingRowColor: MaterialStateProperty.all(
-                  Theme.of(context).colorScheme.background),
+              headingRowColor: WidgetStateProperty.all(
+                  Theme.of(context).colorScheme.surface),
               sortArrowIcon: YaruIcons.pan_up,
               sortAscending: model.sortAscending,
               sortColumnIndex: model.sortColumn.index,
