@@ -6,7 +6,7 @@ import 'package:lxd_x/lxd_x.dart';
 import 'package:provider/provider.dart';
 import 'package:ubuntu_service/ubuntu_service.dart';
 import 'package:wizard_router/wizard_router.dart';
-import 'package:yaru_widgets/yaru_widgets.dart';
+import 'package:yaru/widgets.dart';
 
 import '../remotes/remote_store.dart';
 import 'progress_model.dart';
@@ -121,7 +121,7 @@ class _ProgressViewState extends State<_ProgressView> {
           child: Padding(
             padding: const EdgeInsets.symmetric(horizontal: 24),
             child: YaruBorderContainer(
-              color: Theme.of(context).colorScheme.background,
+              color: Theme.of(context).colorScheme.surface,
               child: Column(
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: [
