@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:provider/provider.dart';
 import 'package:wizard_router/wizard_router.dart';
-import 'package:yaru_widgets/yaru_widgets.dart';
+import 'package:yaru/widgets.dart';
 
 import '../remotes/remote_selector.dart';
 import 'launcher_page.dart';
@@ -20,7 +20,7 @@ class RemoteOsPage extends StatelessWidget {
     return LauncherPage(
       title: Text(l10n.selectOsTitle),
       content: YaruBorderContainer(
-        color: Theme.of(context).colorScheme.background,
+        color: Theme.of(context).colorScheme.surface,
         child: model.images?.when(
           data: (images) => OsSelector(
             items: model.oses,
