@@ -3,7 +3,7 @@ import 'package:context_menu/context_menu.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:terminal_view/terminal_view.dart';
-import 'package:yaru_widgets/yaru_widgets.dart';
+import 'package:yaru/widgets.dart';
 
 import '../settings.dart';
 import 'terminal_actions.dart';
