@@ -1,4 +1,4 @@
-// Mocks generated by Mockito 5.4.0 from annotations
+// Mocks generated by Mockito 5.4.4 from annotations
 // in command_store/test/command_store_test.dart.
 // Do not manually edit this file.
 
@@ -15,6 +15,8 @@ import 'package:mockito/mockito.dart' as _i1;
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
@@ -35,6 +37,7 @@ class MockShortcutStore extends _i1.Mock implements _i2.ShortcutStore {
         Invocation.getter(#hasListeners),
         returnValue: false,
       ) as bool);
+
   @override
   void init(_i3.JSettingsNotifier? settings) => super.noSuchMethod(
         Invocation.method(
@@ -43,6 +46,7 @@ class MockShortcutStore extends _i1.Mock implements _i2.ShortcutStore {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   Set<String> getKeys() => (super.noSuchMethod(
         Invocation.method(
@@ -51,12 +55,14 @@ class MockShortcutStore extends _i1.Mock implements _i2.ShortcutStore {
         ),
         returnValue: <String>{},
       ) as Set<String>);
+
   @override
   List<_i4.LogicalKeySet>? getShortcuts(String? id) =>
       (super.noSuchMethod(Invocation.method(
         #getShortcuts,
         [id],
       )) as List<_i4.LogicalKeySet>?);
+
   @override
   _i5.Future<void> addShortcut(
     String? id,
@@ -73,6 +79,7 @@ class MockShortcutStore extends _i1.Mock implements _i2.ShortcutStore {
         returnValue: _i5.Future<void>.value(),
         returnValueForMissingStub: _i5.Future<void>.value(),
       ) as _i5.Future<void>);
+
   @override
   _i5.Future<void> removeShortcut(
     String? id,
@@ -89,6 +96,7 @@ class MockShortcutStore extends _i1.Mock implements _i2.ShortcutStore {
         returnValue: _i5.Future<void>.value(),
         returnValueForMissingStub: _i5.Future<void>.value(),
       ) as _i5.Future<void>);
+
   @override
   _i5.Future<void> setShortcuts(
     String? id,
@@ -105,6 +113,7 @@ class MockShortcutStore extends _i1.Mock implements _i2.ShortcutStore {
         returnValue: _i5.Future<void>.value(),
         returnValueForMissingStub: _i5.Future<void>.value(),
       ) as _i5.Future<void>);
+
   @override
   _i5.Future<void> removeShortcuts(String? id) => (super.noSuchMethod(
         Invocation.method(
@@ -114,6 +123,7 @@ class MockShortcutStore extends _i1.Mock implements _i2.ShortcutStore {
         returnValue: _i5.Future<void>.value(),
         returnValueForMissingStub: _i5.Future<void>.value(),
       ) as _i5.Future<void>);
+
   @override
   void dispose() => super.noSuchMethod(
         Invocation.method(
@@ -122,6 +132,7 @@ class MockShortcutStore extends _i1.Mock implements _i2.ShortcutStore {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void addListener(_i6.VoidCallback? listener) => super.noSuchMethod(
         Invocation.method(
@@ -130,6 +141,7 @@ class MockShortcutStore extends _i1.Mock implements _i2.ShortcutStore {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void removeListener(_i6.VoidCallback? listener) => super.noSuchMethod(
         Invocation.method(
@@ -138,6 +150,7 @@ class MockShortcutStore extends _i1.Mock implements _i2.ShortcutStore {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void notifyListeners() => super.noSuchMethod(
         Invocation.method(
