@@ -6,7 +6,7 @@ part of 'event.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$_LxdEvent _$$_LxdEventFromJson(Map json) => _$_LxdEvent(
+_$LxdEventImpl _$$LxdEventImplFromJson(Map json) => _$LxdEventImpl(
       type: $enumDecode(_$LxdEventTypeEnumMap, json['type']),
       timestamp: DateTime.parse(json['timestamp'] as String),
       metadata: (json['metadata'] as Map?)?.map(
@@ -16,7 +16,7 @@ _$_LxdEvent _$$_LxdEventFromJson(Map json) => _$_LxdEvent(
       project: json['project'] as String?,
     );
 
-Map<String, dynamic> _$$_LxdEventToJson(_$_LxdEvent instance) =>
+Map<String, dynamic> _$$LxdEventImplToJson(_$LxdEventImpl instance) =>
     <String, dynamic>{
       'type': _$LxdEventTypeEnumMap[instance.type]!,
       'timestamp': instance.timestamp.toIso8601String(),
@@ -31,21 +31,23 @@ const _$LxdEventTypeEnumMap = {
   LxdEventType.lifecycle: 'lifecycle',
 };
 
-_$_LxdEventLogging _$$_LxdEventLoggingFromJson(Map json) => _$_LxdEventLogging(
+_$LxdEventLoggingImpl _$$LxdEventLoggingImplFromJson(Map json) =>
+    _$LxdEventLoggingImpl(
       message: json['message'] as String,
       level: json['level'] as String,
       context: Map<String, String>.from(json['context'] as Map),
     );
 
-Map<String, dynamic> _$$_LxdEventLoggingToJson(_$_LxdEventLogging instance) =>
+Map<String, dynamic> _$$LxdEventLoggingImplToJson(
+        _$LxdEventLoggingImpl instance) =>
     <String, dynamic>{
       'message': instance.message,
       'level': instance.level,
       'context': instance.context,
     };
 
-_$_LxdEventLifecycle _$$_LxdEventLifecycleFromJson(Map json) =>
-    _$_LxdEventLifecycle(
+_$LxdEventLifecycleImpl _$$LxdEventLifecycleImplFromJson(Map json) =>
+    _$LxdEventLifecycleImpl(
       action: json['action'] as String,
       source: json['source'] as String,
       context: (json['context'] as Map?)?.map(
@@ -57,8 +59,8 @@ _$_LxdEventLifecycle _$$_LxdEventLifecycleFromJson(Map json) =>
               Map<String, dynamic>.from(json['requestor'] as Map)),
     );
 
-Map<String, dynamic> _$$_LxdEventLifecycleToJson(
-        _$_LxdEventLifecycle instance) =>
+Map<String, dynamic> _$$LxdEventLifecycleImplToJson(
+        _$LxdEventLifecycleImpl instance) =>
     <String, dynamic>{
       'action': instance.action,
       'source': instance.source,
@@ -66,16 +68,16 @@ Map<String, dynamic> _$$_LxdEventLifecycleToJson(
       'requestor': instance.requestor?.toJson(),
     };
 
-_$_LxdEventLifecycleRequestor _$$_LxdEventLifecycleRequestorFromJson(
+_$LxdEventLifecycleRequestorImpl _$$LxdEventLifecycleRequestorImplFromJson(
         Map json) =>
-    _$_LxdEventLifecycleRequestor(
+    _$LxdEventLifecycleRequestorImpl(
       username: json['username'] as String,
       protocol: json['protocol'] as String,
       address: json['address'] as String,
     );
 
-Map<String, dynamic> _$$_LxdEventLifecycleRequestorToJson(
-        _$_LxdEventLifecycleRequestor instance) =>
+Map<String, dynamic> _$$LxdEventLifecycleRequestorImplToJson(
+        _$LxdEventLifecycleRequestorImpl instance) =>
     <String, dynamic>{
       'username': instance.username,
       'protocol': instance.protocol,
