@@ -12,7 +12,7 @@ part of 'profile.dart';
 T _$identity<T>(T value) => value;
 
 final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
 
 LxdProfile _$LxdProfileFromJson(Map<String, dynamic> json) {
   return _LxdProfile.fromJson(json);
@@ -24,8 +24,12 @@ mixin _$LxdProfile {
   String get description => throw _privateConstructorUsedError;
   String get name => throw _privateConstructorUsedError;
 
+  /// Serializes this LxdProfile to a JSON map.
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
-  @JsonKey(ignore: true)
+
+  /// Create a copy of LxdProfile
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
   $LxdProfileCopyWith<LxdProfile> get copyWith =>
       throw _privateConstructorUsedError;
 }
@@ -49,6 +53,8 @@ class _$LxdProfileCopyWithImpl<$Res, $Val extends LxdProfile>
   // ignore: unused_field
   final $Res Function($Val) _then;
 
+  /// Create a copy of LxdProfile
+  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -74,24 +80,26 @@ class _$LxdProfileCopyWithImpl<$Res, $Val extends LxdProfile>
 }
 
 /// @nodoc
-abstract class _$$_LxdProfileCopyWith<$Res>
+abstract class _$$LxdProfileImplCopyWith<$Res>
     implements $LxdProfileCopyWith<$Res> {
-  factory _$$_LxdProfileCopyWith(
-          _$_LxdProfile value, $Res Function(_$_LxdProfile) then) =
-      __$$_LxdProfileCopyWithImpl<$Res>;
+  factory _$$LxdProfileImplCopyWith(
+          _$LxdProfileImpl value, $Res Function(_$LxdProfileImpl) then) =
+      __$$LxdProfileImplCopyWithImpl<$Res>;
   @override
   @useResult
   $Res call({Map<String, String> config, String description, String name});
 }
 
 /// @nodoc
-class __$$_LxdProfileCopyWithImpl<$Res>
-    extends _$LxdProfileCopyWithImpl<$Res, _$_LxdProfile>
-    implements _$$_LxdProfileCopyWith<$Res> {
-  __$$_LxdProfileCopyWithImpl(
-      _$_LxdProfile _value, $Res Function(_$_LxdProfile) _then)
+class __$$LxdProfileImplCopyWithImpl<$Res>
+    extends _$LxdProfileCopyWithImpl<$Res, _$LxdProfileImpl>
+    implements _$$LxdProfileImplCopyWith<$Res> {
+  __$$LxdProfileImplCopyWithImpl(
+      _$LxdProfileImpl _value, $Res Function(_$LxdProfileImpl) _then)
       : super(_value, _then);
 
+  /// Create a copy of LxdProfile
+  /// with the given fields replaced by the non-null parameter values.
   @pragma('vm:prefer-inline')
   @override
   $Res call({
@@ -99,7 +107,7 @@ class __$$_LxdProfileCopyWithImpl<$Res>
     Object? description = null,
     Object? name = null,
   }) {
-    return _then(_$_LxdProfile(
+    return _then(_$LxdProfileImpl(
       config: null == config
           ? _value._config
           : config // ignore: cast_nullable_to_non_nullable
@@ -118,15 +126,15 @@ class __$$_LxdProfileCopyWithImpl<$Res>
 
 /// @nodoc
 @JsonSerializable()
-class _$_LxdProfile implements _LxdProfile {
-  const _$_LxdProfile(
+class _$LxdProfileImpl implements _LxdProfile {
+  const _$LxdProfileImpl(
       {required final Map<String, String> config,
       required this.description,
       required this.name})
       : _config = config;
 
-  factory _$_LxdProfile.fromJson(Map<String, dynamic> json) =>
-      _$$_LxdProfileFromJson(json);
+  factory _$LxdProfileImpl.fromJson(Map<String, dynamic> json) =>
+      _$$LxdProfileImplFromJson(json);
 
   final Map<String, String> _config;
   @override
@@ -147,30 +155,32 @@ class _$_LxdProfile implements _LxdProfile {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
-            other is _$_LxdProfile &&
+            other is _$LxdProfileImpl &&
             const DeepCollectionEquality().equals(other._config, _config) &&
             (identical(other.description, description) ||
                 other.description == description) &&
             (identical(other.name, name) || other.name == name));
   }
 
-  @JsonKey(ignore: true)
+  @JsonKey(includeFromJson: false, includeToJson: false)
   @override
   int get hashCode => Object.hash(runtimeType,
       const DeepCollectionEquality().hash(_config), description, name);
 
-  @JsonKey(ignore: true)
+  /// Create a copy of LxdProfile
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
   @override
   @pragma('vm:prefer-inline')
-  _$$_LxdProfileCopyWith<_$_LxdProfile> get copyWith =>
-      __$$_LxdProfileCopyWithImpl<_$_LxdProfile>(this, _$identity);
+  _$$LxdProfileImplCopyWith<_$LxdProfileImpl> get copyWith =>
+      __$$LxdProfileImplCopyWithImpl<_$LxdProfileImpl>(this, _$identity);
 
   @override
   Map<String, dynamic> toJson() {
-    return _$$_LxdProfileToJson(
+    return _$$LxdProfileImplToJson(
       this,
     );
   }
@@ -180,10 +190,10 @@ abstract class _LxdProfile implements LxdProfile {
   const factory _LxdProfile(
       {required final Map<String, String> config,
       required final String description,
-      required final String name}) = _$_LxdProfile;
+      required final String name}) = _$LxdProfileImpl;
 
   factory _LxdProfile.fromJson(Map<String, dynamic> json) =
-      _$_LxdProfile.fromJson;
+      _$LxdProfileImpl.fromJson;
 
   @override
   Map<String, String> get config;
@@ -191,8 +201,11 @@ abstract class _LxdProfile implements LxdProfile {
   String get description;
   @override
   String get name;
+
+  /// Create a copy of LxdProfile
+  /// with the given fields replaced by the non-null parameter values.
   @override
-  @JsonKey(ignore: true)
-  _$$_LxdProfileCopyWith<_$_LxdProfile> get copyWith =>
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  _$$LxdProfileImplCopyWith<_$LxdProfileImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }
