@@ -6,7 +6,7 @@ part of 'project.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$_LxdProject _$$_LxdProjectFromJson(Map json) => _$_LxdProject(
+_$LxdProjectImpl _$$LxdProjectImplFromJson(Map json) => _$LxdProjectImpl(
       config: Map<String, String>.from(json['config'] as Map),
       description: json['description'] as String,
       name: json['name'] as String,
@@ -14,7 +14,7 @@ _$_LxdProject _$$_LxdProjectFromJson(Map json) => _$_LxdProject(
           (json['used_by'] as List<dynamic>).map((e) => e as String).toList(),
     );
 
-Map<String, dynamic> _$$_LxdProjectToJson(_$_LxdProject instance) =>
+Map<String, dynamic> _$$LxdProjectImplToJson(_$LxdProjectImpl instance) =>
     <String, dynamic>{
       'config': instance.config,
       'description': instance.description,
