@@ -6,14 +6,16 @@ part of 'response.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$LxdSyncResponse _$$LxdSyncResponseFromJson(Map json) => _$LxdSyncResponse(
+_$LxdSyncResponseImpl _$$LxdSyncResponseImplFromJson(Map json) =>
+    _$LxdSyncResponseImpl(
       status: json['status'] as String,
-      statusCode: json['status_code'] as int,
+      statusCode: (json['status_code'] as num).toInt(),
       metadata: json['metadata'],
       $type: json['type'] as String?,
     );
 
-Map<String, dynamic> _$$LxdSyncResponseToJson(_$LxdSyncResponse instance) =>
+Map<String, dynamic> _$$LxdSyncResponseImplToJson(
+        _$LxdSyncResponseImpl instance) =>
     <String, dynamic>{
       'status': instance.status,
       'status_code': instance.statusCode,
@@ -21,15 +23,17 @@ Map<String, dynamic> _$$LxdSyncResponseToJson(_$LxdSyncResponse instance) =>
       'type': instance.$type,
     };
 
-_$LxdAsyncResponse _$$LxdAsyncResponseFromJson(Map json) => _$LxdAsyncResponse(
+_$LxdAsyncResponseImpl _$$LxdAsyncResponseImplFromJson(Map json) =>
+    _$LxdAsyncResponseImpl(
       status: json['status'] as String,
-      statusCode: json['status_code'] as int,
+      statusCode: (json['status_code'] as num).toInt(),
       operation: json['operation'] as String,
       metadata: Map<String, dynamic>.from(json['metadata'] as Map),
       $type: json['type'] as String?,
     );
 
-Map<String, dynamic> _$$LxdAsyncResponseToJson(_$LxdAsyncResponse instance) =>
+Map<String, dynamic> _$$LxdAsyncResponseImplToJson(
+        _$LxdAsyncResponseImpl instance) =>
     <String, dynamic>{
       'status': instance.status,
       'status_code': instance.statusCode,
@@ -38,16 +42,18 @@ Map<String, dynamic> _$$LxdAsyncResponseToJson(_$LxdAsyncResponse instance) =>
       'type': instance.$type,
     };
 
-_$LxdErrorResponse _$$LxdErrorResponseFromJson(Map json) => _$LxdErrorResponse(
+_$LxdErrorResponseImpl _$$LxdErrorResponseImplFromJson(Map json) =>
+    _$LxdErrorResponseImpl(
       error: json['error'] as String,
-      errorCode: json['error_code'] as int,
+      errorCode: (json['error_code'] as num).toInt(),
       metadata: (json['metadata'] as Map?)?.map(
         (k, e) => MapEntry(k as String, e),
       ),
       $type: json['type'] as String?,
     );
 
-Map<String, dynamic> _$$LxdErrorResponseToJson(_$LxdErrorResponse instance) =>
+Map<String, dynamic> _$$LxdErrorResponseImplToJson(
+        _$LxdErrorResponseImpl instance) =>
     <String, dynamic>{
       'error': instance.error,
       'error_code': instance.errorCode,
