@@ -1,4 +1,4 @@
-// Mocks generated by Mockito 5.4.0 from annotations
+// Mocks generated by Mockito 5.4.4 from annotations
 // in lxd/test/lxd_http.dart.
 // Do not manually edit this file.
 
@@ -8,11 +8,14 @@ import 'dart:convert' as _i3;
 import 'dart:io' as _i2;
 
 import 'package:mockito/mockito.dart' as _i1;
+import 'package:mockito/src/dummies.dart' as _i5;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
@@ -129,6 +132,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
           Invocation.getter(#idleTimeout),
         ),
       ) as Duration);
+
   @override
   set idleTimeout(Duration? _idleTimeout) => super.noSuchMethod(
         Invocation.setter(
@@ -137,6 +141,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   set connectionTimeout(Duration? _connectionTimeout) => super.noSuchMethod(
         Invocation.setter(
@@ -145,6 +150,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   set maxConnectionsPerHost(int? _maxConnectionsPerHost) => super.noSuchMethod(
         Invocation.setter(
@@ -153,11 +159,13 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   bool get autoUncompress => (super.noSuchMethod(
         Invocation.getter(#autoUncompress),
         returnValue: false,
       ) as bool);
+
   @override
   set autoUncompress(bool? _autoUncompress) => super.noSuchMethod(
         Invocation.setter(
@@ -166,6 +174,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   set userAgent(String? _userAgent) => super.noSuchMethod(
         Invocation.setter(
@@ -174,6 +183,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   set authenticate(
           _i4.Future<bool> Function(
@@ -188,6 +198,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   set connectionFactory(
           _i4.Future<_i2.ConnectionTask<_i2.Socket>> Function(
@@ -202,6 +213,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   set findProxy(String Function(Uri)? f) => super.noSuchMethod(
         Invocation.setter(
@@ -210,6 +222,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   set authenticateProxy(
           _i4.Future<bool> Function(
@@ -225,6 +238,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   set badCertificateCallback(
           bool Function(
@@ -239,6 +253,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   set keyLog(dynamic Function(String)? callback) => super.noSuchMethod(
         Invocation.setter(
@@ -247,6 +262,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i4.Future<_i2.HttpClientRequest> open(
     String? method,
@@ -278,6 +294,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
           ),
         )),
       ) as _i4.Future<_i2.HttpClientRequest>);
+
   @override
   _i4.Future<_i2.HttpClientRequest> openUrl(
     String? method,
@@ -303,6 +320,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
           ),
         )),
       ) as _i4.Future<_i2.HttpClientRequest>);
+
   @override
   _i4.Future<_i2.HttpClientRequest> get(
     String? host,
@@ -331,6 +349,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
           ),
         )),
       ) as _i4.Future<_i2.HttpClientRequest>);
+
   @override
   _i4.Future<_i2.HttpClientRequest> getUrl(Uri? url) => (super.noSuchMethod(
         Invocation.method(
@@ -346,6 +365,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
           ),
         )),
       ) as _i4.Future<_i2.HttpClientRequest>);
+
   @override
   _i4.Future<_i2.HttpClientRequest> post(
     String? host,
@@ -374,6 +394,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
           ),
         )),
       ) as _i4.Future<_i2.HttpClientRequest>);
+
   @override
   _i4.Future<_i2.HttpClientRequest> postUrl(Uri? url) => (super.noSuchMethod(
         Invocation.method(
@@ -389,6 +410,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
           ),
         )),
       ) as _i4.Future<_i2.HttpClientRequest>);
+
   @override
   _i4.Future<_i2.HttpClientRequest> put(
     String? host,
@@ -417,6 +439,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
           ),
         )),
       ) as _i4.Future<_i2.HttpClientRequest>);
+
   @override
   _i4.Future<_i2.HttpClientRequest> putUrl(Uri? url) => (super.noSuchMethod(
         Invocation.method(
@@ -432,6 +455,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
           ),
         )),
       ) as _i4.Future<_i2.HttpClientRequest>);
+
   @override
   _i4.Future<_i2.HttpClientRequest> delete(
     String? host,
@@ -460,6 +484,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
           ),
         )),
       ) as _i4.Future<_i2.HttpClientRequest>);
+
   @override
   _i4.Future<_i2.HttpClientRequest> deleteUrl(Uri? url) => (super.noSuchMethod(
         Invocation.method(
@@ -475,6 +500,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
           ),
         )),
       ) as _i4.Future<_i2.HttpClientRequest>);
+
   @override
   _i4.Future<_i2.HttpClientRequest> patch(
     String? host,
@@ -503,6 +529,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
           ),
         )),
       ) as _i4.Future<_i2.HttpClientRequest>);
+
   @override
   _i4.Future<_i2.HttpClientRequest> patchUrl(Uri? url) => (super.noSuchMethod(
         Invocation.method(
@@ -518,6 +545,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
           ),
         )),
       ) as _i4.Future<_i2.HttpClientRequest>);
+
   @override
   _i4.Future<_i2.HttpClientRequest> head(
     String? host,
@@ -546,6 +574,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
           ),
         )),
       ) as _i4.Future<_i2.HttpClientRequest>);
+
   @override
   _i4.Future<_i2.HttpClientRequest> headUrl(Uri? url) => (super.noSuchMethod(
         Invocation.method(
@@ -561,6 +590,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
           ),
         )),
       ) as _i4.Future<_i2.HttpClientRequest>);
+
   @override
   void addCredentials(
     Uri? url,
@@ -578,6 +608,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void addProxyCredentials(
     String? host,
@@ -597,6 +628,7 @@ class MockHttpClient extends _i1.Mock implements _i2.HttpClient {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void close({bool? force = false}) => super.noSuchMethod(
         Invocation.method(
@@ -621,6 +653,7 @@ class MockHttpClientRequest extends _i1.Mock implements _i2.HttpClientRequest {
         Invocation.getter(#persistentConnection),
         returnValue: false,
       ) as bool);
+
   @override
   set persistentConnection(bool? _persistentConnection) => super.noSuchMethod(
         Invocation.setter(
@@ -629,11 +662,13 @@ class MockHttpClientRequest extends _i1.Mock implements _i2.HttpClientRequest {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   bool get followRedirects => (super.noSuchMethod(
         Invocation.getter(#followRedirects),
         returnValue: false,
       ) as bool);
+
   @override
   set followRedirects(bool? _followRedirects) => super.noSuchMethod(
         Invocation.setter(
@@ -642,11 +677,13 @@ class MockHttpClientRequest extends _i1.Mock implements _i2.HttpClientRequest {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   int get maxRedirects => (super.noSuchMethod(
         Invocation.getter(#maxRedirects),
         returnValue: 0,
       ) as int);
+
   @override
   set maxRedirects(int? _maxRedirects) => super.noSuchMethod(
         Invocation.setter(
@@ -655,11 +692,13 @@ class MockHttpClientRequest extends _i1.Mock implements _i2.HttpClientRequest {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   int get contentLength => (super.noSuchMethod(
         Invocation.getter(#contentLength),
         returnValue: 0,
       ) as int);
+
   @override
   set contentLength(int? _contentLength) => super.noSuchMethod(
         Invocation.setter(
@@ -668,11 +707,13 @@ class MockHttpClientRequest extends _i1.Mock implements _i2.HttpClientRequest {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   bool get bufferOutput => (super.noSuchMethod(
         Invocation.getter(#bufferOutput),
         returnValue: false,
       ) as bool);
+
   @override
   set bufferOutput(bool? _bufferOutput) => super.noSuchMethod(
         Invocation.setter(
@@ -681,11 +722,16 @@ class MockHttpClientRequest extends _i1.Mock implements _i2.HttpClientRequest {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   String get method => (super.noSuchMethod(
         Invocation.getter(#method),
-        returnValue: '',
+        returnValue: _i5.dummyValue<String>(
+          this,
+          Invocation.getter(#method),
+        ),
       ) as String);
+
   @override
   Uri get uri => (super.noSuchMethod(
         Invocation.getter(#uri),
@@ -694,6 +740,7 @@ class MockHttpClientRequest extends _i1.Mock implements _i2.HttpClientRequest {
           Invocation.getter(#uri),
         ),
       ) as Uri);
+
   @override
   _i2.HttpHeaders get headers => (super.noSuchMethod(
         Invocation.getter(#headers),
@@ -702,11 +749,13 @@ class MockHttpClientRequest extends _i1.Mock implements _i2.HttpClientRequest {
           Invocation.getter(#headers),
         ),
       ) as _i2.HttpHeaders);
+
   @override
   List<_i2.Cookie> get cookies => (super.noSuchMethod(
         Invocation.getter(#cookies),
         returnValue: <_i2.Cookie>[],
       ) as List<_i2.Cookie>);
+
   @override
   _i4.Future<_i2.HttpClientResponse> get done => (super.noSuchMethod(
         Invocation.getter(#done),
@@ -716,6 +765,7 @@ class MockHttpClientRequest extends _i1.Mock implements _i2.HttpClientRequest {
           Invocation.getter(#done),
         )),
       ) as _i4.Future<_i2.HttpClientResponse>);
+
   @override
   _i3.Encoding get encoding => (super.noSuchMethod(
         Invocation.getter(#encoding),
@@ -724,6 +774,7 @@ class MockHttpClientRequest extends _i1.Mock implements _i2.HttpClientRequest {
           Invocation.getter(#encoding),
         ),
       ) as _i3.Encoding);
+
   @override
   set encoding(_i3.Encoding? _encoding) => super.noSuchMethod(
         Invocation.setter(
@@ -732,6 +783,7 @@ class MockHttpClientRequest extends _i1.Mock implements _i2.HttpClientRequest {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i4.Future<_i2.HttpClientResponse> close() => (super.noSuchMethod(
         Invocation.method(
@@ -747,6 +799,7 @@ class MockHttpClientRequest extends _i1.Mock implements _i2.HttpClientRequest {
           ),
         )),
       ) as _i4.Future<_i2.HttpClientResponse>);
+
   @override
   void abort([
     Object? exception,
@@ -762,6 +815,7 @@ class MockHttpClientRequest extends _i1.Mock implements _i2.HttpClientRequest {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void add(List<int>? data) => super.noSuchMethod(
         Invocation.method(
@@ -770,6 +824,7 @@ class MockHttpClientRequest extends _i1.Mock implements _i2.HttpClientRequest {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void write(Object? object) => super.noSuchMethod(
         Invocation.method(
@@ -778,6 +833,7 @@ class MockHttpClientRequest extends _i1.Mock implements _i2.HttpClientRequest {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void writeAll(
     Iterable<dynamic>? objects, [
@@ -793,6 +849,7 @@ class MockHttpClientRequest extends _i1.Mock implements _i2.HttpClientRequest {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void writeln([Object? object = r'']) => super.noSuchMethod(
         Invocation.method(
@@ -801,6 +858,7 @@ class MockHttpClientRequest extends _i1.Mock implements _i2.HttpClientRequest {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void writeCharCode(int? charCode) => super.noSuchMethod(
         Invocation.method(
@@ -809,6 +867,7 @@ class MockHttpClientRequest extends _i1.Mock implements _i2.HttpClientRequest {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void addError(
     Object? error, [
@@ -824,6 +883,7 @@ class MockHttpClientRequest extends _i1.Mock implements _i2.HttpClientRequest {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i4.Future<dynamic> addStream(_i4.Stream<List<int>>? stream) =>
       (super.noSuchMethod(
@@ -833,6 +893,7 @@ class MockHttpClientRequest extends _i1.Mock implements _i2.HttpClientRequest {
         ),
         returnValue: _i4.Future<dynamic>.value(),
       ) as _i4.Future<dynamic>);
+
   @override
   _i4.Future<dynamic> flush() => (super.noSuchMethod(
         Invocation.method(
@@ -857,37 +918,47 @@ class MockHttpClientResponse extends _i1.Mock
         Invocation.getter(#statusCode),
         returnValue: 0,
       ) as int);
+
   @override
   String get reasonPhrase => (super.noSuchMethod(
         Invocation.getter(#reasonPhrase),
-        returnValue: '',
+        returnValue: _i5.dummyValue<String>(
+          this,
+          Invocation.getter(#reasonPhrase),
+        ),
       ) as String);
+
   @override
   int get contentLength => (super.noSuchMethod(
         Invocation.getter(#contentLength),
         returnValue: 0,
       ) as int);
+
   @override
   _i2.HttpClientResponseCompressionState get compressionState =>
       (super.noSuchMethod(
         Invocation.getter(#compressionState),
         returnValue: _i2.HttpClientResponseCompressionState.notCompressed,
       ) as _i2.HttpClientResponseCompressionState);
+
   @override
   bool get persistentConnection => (super.noSuchMethod(
         Invocation.getter(#persistentConnection),
         returnValue: false,
       ) as bool);
+
   @override
   bool get isRedirect => (super.noSuchMethod(
         Invocation.getter(#isRedirect),
         returnValue: false,
       ) as bool);
+
   @override
   List<_i2.RedirectInfo> get redirects => (super.noSuchMethod(
         Invocation.getter(#redirects),
         returnValue: <_i2.RedirectInfo>[],
       ) as List<_i2.RedirectInfo>);
+
   @override
   _i2.HttpHeaders get headers => (super.noSuchMethod(
         Invocation.getter(#headers),
@@ -896,41 +967,49 @@ class MockHttpClientResponse extends _i1.Mock
           Invocation.getter(#headers),
         ),
       ) as _i2.HttpHeaders);
+
   @override
   List<_i2.Cookie> get cookies => (super.noSuchMethod(
         Invocation.getter(#cookies),
         returnValue: <_i2.Cookie>[],
       ) as List<_i2.Cookie>);
+
   @override
   bool get isBroadcast => (super.noSuchMethod(
         Invocation.getter(#isBroadcast),
         returnValue: false,
       ) as bool);
+
   @override
   _i4.Future<int> get length => (super.noSuchMethod(
         Invocation.getter(#length),
         returnValue: _i4.Future<int>.value(0),
       ) as _i4.Future<int>);
+
   @override
   _i4.Future<bool> get isEmpty => (super.noSuchMethod(
         Invocation.getter(#isEmpty),
         returnValue: _i4.Future<bool>.value(false),
       ) as _i4.Future<bool>);
+
   @override
   _i4.Future<List<int>> get first => (super.noSuchMethod(
         Invocation.getter(#first),
         returnValue: _i4.Future<List<int>>.value(<int>[]),
       ) as _i4.Future<List<int>>);
+
   @override
   _i4.Future<List<int>> get last => (super.noSuchMethod(
         Invocation.getter(#last),
         returnValue: _i4.Future<List<int>>.value(<int>[]),
       ) as _i4.Future<List<int>>);
+
   @override
   _i4.Future<List<int>> get single => (super.noSuchMethod(
         Invocation.getter(#single),
         returnValue: _i4.Future<List<int>>.value(<int>[]),
       ) as _i4.Future<List<int>>);
+
   @override
   _i4.Future<_i2.HttpClientResponse> redirect([
     String? method,
@@ -959,6 +1038,7 @@ class MockHttpClientResponse extends _i1.Mock
           ),
         )),
       ) as _i4.Future<_i2.HttpClientResponse>);
+
   @override
   _i4.Future<_i2.Socket> detachSocket() => (super.noSuchMethod(
         Invocation.method(
@@ -973,6 +1053,7 @@ class MockHttpClientResponse extends _i1.Mock
           ),
         )),
       ) as _i4.Future<_i2.Socket>);
+
   @override
   _i4.Stream<List<int>> asBroadcastStream({
     void Function(_i4.StreamSubscription<List<int>>)? onListen,
@@ -989,6 +1070,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Stream<List<int>>.empty(),
       ) as _i4.Stream<List<int>>);
+
   @override
   _i4.StreamSubscription<List<int>> listen(
     void Function(List<int>)? onData, {
@@ -1019,6 +1101,7 @@ class MockHttpClientResponse extends _i1.Mock
           ),
         ),
       ) as _i4.StreamSubscription<List<int>>);
+
   @override
   _i4.Stream<List<int>> where(bool Function(List<int>)? test) =>
       (super.noSuchMethod(
@@ -1028,6 +1111,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Stream<List<int>>.empty(),
       ) as _i4.Stream<List<int>>);
+
   @override
   _i4.Stream<S> map<S>(S Function(List<int>)? convert) => (super.noSuchMethod(
         Invocation.method(
@@ -1036,6 +1120,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Stream<S>.empty(),
       ) as _i4.Stream<S>);
+
   @override
   _i4.Stream<E> asyncMap<E>(_i4.FutureOr<E> Function(List<int>)? convert) =>
       (super.noSuchMethod(
@@ -1045,6 +1130,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Stream<E>.empty(),
       ) as _i4.Stream<E>);
+
   @override
   _i4.Stream<E> asyncExpand<E>(_i4.Stream<E>? Function(List<int>)? convert) =>
       (super.noSuchMethod(
@@ -1054,6 +1140,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Stream<E>.empty(),
       ) as _i4.Stream<E>);
+
   @override
   _i4.Stream<List<int>> handleError(
     Function? onError, {
@@ -1067,6 +1154,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Stream<List<int>>.empty(),
       ) as _i4.Stream<List<int>>);
+
   @override
   _i4.Stream<S> expand<S>(Iterable<S> Function(List<int>)? convert) =>
       (super.noSuchMethod(
@@ -1076,6 +1164,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Stream<S>.empty(),
       ) as _i4.Stream<S>);
+
   @override
   _i4.Future<dynamic> pipe(_i4.StreamConsumer<List<int>>? streamConsumer) =>
       (super.noSuchMethod(
@@ -1085,6 +1174,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Future<dynamic>.value(),
       ) as _i4.Future<dynamic>);
+
   @override
   _i4.Stream<S> transform<S>(
           _i4.StreamTransformer<List<int>, S>? streamTransformer) =>
@@ -1095,6 +1185,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Stream<S>.empty(),
       ) as _i4.Stream<S>);
+
   @override
   _i4.Future<List<int>> reduce(
           List<int> Function(
@@ -1108,6 +1199,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Future<List<int>>.value(<int>[]),
       ) as _i4.Future<List<int>>);
+
   @override
   _i4.Future<S> fold<S>(
     S? initialValue,
@@ -1124,25 +1216,46 @@ class MockHttpClientResponse extends _i1.Mock
             combine,
           ],
         ),
-        returnValue: _FakeFuture_8<S>(
-          this,
-          Invocation.method(
-            #fold,
-            [
-              initialValue,
-              combine,
-            ],
-          ),
-        ),
+        returnValue: _i5.ifNotNull(
+              _i5.dummyValueOrNull<S>(
+                this,
+                Invocation.method(
+                  #fold,
+                  [
+                    initialValue,
+                    combine,
+                  ],
+                ),
+              ),
+              (S v) => _i4.Future<S>.value(v),
+            ) ??
+            _FakeFuture_8<S>(
+              this,
+              Invocation.method(
+                #fold,
+                [
+                  initialValue,
+                  combine,
+                ],
+              ),
+            ),
       ) as _i4.Future<S>);
+
   @override
   _i4.Future<String> join([String? separator = r'']) => (super.noSuchMethod(
         Invocation.method(
           #join,
           [separator],
         ),
-        returnValue: _i4.Future<String>.value(''),
+        returnValue: _i4.Future<String>.value(_i5.dummyValue<String>(
+          this,
+          Invocation.method(
+            #join,
+            [separator],
+          ),
+        )),
       ) as _i4.Future<String>);
+
   @override
   _i4.Future<bool> contains(Object? needle) => (super.noSuchMethod(
         Invocation.method(
@@ -1151,6 +1264,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Future<bool>.value(false),
       ) as _i4.Future<bool>);
+
   @override
   _i4.Future<void> forEach(void Function(List<int>)? action) =>
       (super.noSuchMethod(
@@ -1161,6 +1275,7 @@ class MockHttpClientResponse extends _i1.Mock
         returnValue: _i4.Future<void>.value(),
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
+
   @override
   _i4.Future<bool> every(bool Function(List<int>)? test) => (super.noSuchMethod(
         Invocation.method(
@@ -1169,6 +1284,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Future<bool>.value(false),
       ) as _i4.Future<bool>);
+
   @override
   _i4.Future<bool> any(bool Function(List<int>)? test) => (super.noSuchMethod(
         Invocation.method(
@@ -1177,6 +1293,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Future<bool>.value(false),
       ) as _i4.Future<bool>);
+
   @override
   _i4.Stream<R> cast<R>() => (super.noSuchMethod(
         Invocation.method(
@@ -1185,6 +1302,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Stream<R>.empty(),
       ) as _i4.Stream<R>);
+
   @override
   _i4.Future<List<List<int>>> toList() => (super.noSuchMethod(
         Invocation.method(
@@ -1193,6 +1311,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Future<List<List<int>>>.value(<List<int>>[]),
       ) as _i4.Future<List<List<int>>>);
+
   @override
   _i4.Future<Set<List<int>>> toSet() => (super.noSuchMethod(
         Invocation.method(
@@ -1201,20 +1320,32 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Future<Set<List<int>>>.value(<List<int>>{}),
       ) as _i4.Future<Set<List<int>>>);
+
   @override
   _i4.Future<E> drain<E>([E? futureValue]) => (super.noSuchMethod(
         Invocation.method(
           #drain,
           [futureValue],
         ),
-        returnValue: _FakeFuture_8<E>(
-          this,
-          Invocation.method(
-            #drain,
-            [futureValue],
-          ),
-        ),
+        returnValue: _i5.ifNotNull(
+              _i5.dummyValueOrNull<E>(
+                this,
+                Invocation.method(
+                  #drain,
+                  [futureValue],
+                ),
+              ),
+              (E v) => _i4.Future<E>.value(v),
+            ) ??
+            _FakeFuture_8<E>(
+              this,
+              Invocation.method(
+                #drain,
+                [futureValue],
+              ),
+            ),
       ) as _i4.Future<E>);
+
   @override
   _i4.Stream<List<int>> take(int? count) => (super.noSuchMethod(
         Invocation.method(
@@ -1223,6 +1354,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Stream<List<int>>.empty(),
       ) as _i4.Stream<List<int>>);
+
   @override
   _i4.Stream<List<int>> takeWhile(bool Function(List<int>)? test) =>
       (super.noSuchMethod(
@@ -1232,6 +1364,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Stream<List<int>>.empty(),
       ) as _i4.Stream<List<int>>);
+
   @override
   _i4.Stream<List<int>> skip(int? count) => (super.noSuchMethod(
         Invocation.method(
@@ -1240,6 +1373,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Stream<List<int>>.empty(),
       ) as _i4.Stream<List<int>>);
+
   @override
   _i4.Stream<List<int>> skipWhile(bool Function(List<int>)? test) =>
       (super.noSuchMethod(
@@ -1249,6 +1383,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Stream<List<int>>.empty(),
       ) as _i4.Stream<List<int>>);
+
   @override
   _i4.Stream<List<int>> distinct(
           [bool Function(
@@ -1262,6 +1397,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Stream<List<int>>.empty(),
       ) as _i4.Stream<List<int>>);
+
   @override
   _i4.Future<List<int>> firstWhere(
     bool Function(List<int>)? test, {
@@ -1275,6 +1411,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Future<List<int>>.value(<int>[]),
       ) as _i4.Future<List<int>>);
+
   @override
   _i4.Future<List<int>> lastWhere(
     bool Function(List<int>)? test, {
@@ -1288,6 +1425,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Future<List<int>>.value(<int>[]),
       ) as _i4.Future<List<int>>);
+
   @override
   _i4.Future<List<int>> singleWhere(
     bool Function(List<int>)? test, {
@@ -1301,6 +1439,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Future<List<int>>.value(<int>[]),
       ) as _i4.Future<List<int>>);
+
   @override
   _i4.Future<List<int>> elementAt(int? index) => (super.noSuchMethod(
         Invocation.method(
@@ -1309,6 +1448,7 @@ class MockHttpClientResponse extends _i1.Mock
         ),
         returnValue: _i4.Future<List<int>>.value(<int>[]),
       ) as _i4.Future<List<int>>);
+
   @override
   _i4.Stream<List<int>> timeout(
     Duration? timeLimit, {
@@ -1340,6 +1480,7 @@ class MockHttpHeaders extends _i1.Mock implements _i2.HttpHeaders {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   set expires(DateTime? _expires) => super.noSuchMethod(
         Invocation.setter(
@@ -1348,6 +1489,7 @@ class MockHttpHeaders extends _i1.Mock implements _i2.HttpHeaders {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   set ifModifiedSince(DateTime? _ifModifiedSince) => super.noSuchMethod(
         Invocation.setter(
@@ -1356,6 +1498,7 @@ class MockHttpHeaders extends _i1.Mock implements _i2.HttpHeaders {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   set host(String? _host) => super.noSuchMethod(
         Invocation.setter(
@@ -1364,6 +1507,7 @@ class MockHttpHeaders extends _i1.Mock implements _i2.HttpHeaders {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   set port(int? _port) => super.noSuchMethod(
         Invocation.setter(
@@ -1372,6 +1516,7 @@ class MockHttpHeaders extends _i1.Mock implements _i2.HttpHeaders {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   set contentType(_i2.ContentType? _contentType) => super.noSuchMethod(
         Invocation.setter(
@@ -1380,11 +1525,13 @@ class MockHttpHeaders extends _i1.Mock implements _i2.HttpHeaders {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   int get contentLength => (super.noSuchMethod(
         Invocation.getter(#contentLength),
         returnValue: 0,
       ) as int);
+
   @override
   set contentLength(int? _contentLength) => super.noSuchMethod(
         Invocation.setter(
@@ -1393,11 +1540,13 @@ class MockHttpHeaders extends _i1.Mock implements _i2.HttpHeaders {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   bool get persistentConnection => (super.noSuchMethod(
         Invocation.getter(#persistentConnection),
         returnValue: false,
       ) as bool);
+
   @override
   set persistentConnection(bool? _persistentConnection) => super.noSuchMethod(
         Invocation.setter(
@@ -1406,11 +1555,13 @@ class MockHttpHeaders extends _i1.Mock implements _i2.HttpHeaders {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   bool get chunkedTransferEncoding => (super.noSuchMethod(
         Invocation.getter(#chunkedTransferEncoding),
         returnValue: false,
       ) as bool);
+
   @override
   set chunkedTransferEncoding(bool? _chunkedTransferEncoding) =>
       super.noSuchMethod(
@@ -1420,17 +1571,20 @@ class MockHttpHeaders extends _i1.Mock implements _i2.HttpHeaders {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   List<String>? operator [](String? name) =>
       (super.noSuchMethod(Invocation.method(
         #[],
         [name],
       )) as List<String>?);
+
   @override
   String? value(String? name) => (super.noSuchMethod(Invocation.method(
         #value,
         [name],
       )) as String?);
+
   @override
   void add(
     String? name,
@@ -1448,6 +1602,7 @@ class MockHttpHeaders extends _i1.Mock implements _i2.HttpHeaders {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void set(
     String? name,
@@ -1465,6 +1620,7 @@ class MockHttpHeaders extends _i1.Mock implements _i2.HttpHeaders {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void remove(
     String? name,
@@ -1480,6 +1636,7 @@ class MockHttpHeaders extends _i1.Mock implements _i2.HttpHeaders {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void removeAll(String? name) => super.noSuchMethod(
         Invocation.method(
@@ -1488,6 +1645,7 @@ class MockHttpHeaders extends _i1.Mock implements _i2.HttpHeaders {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void forEach(
           void Function(
@@ -1501,6 +1659,7 @@ class MockHttpHeaders extends _i1.Mock implements _i2.HttpHeaders {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void noFolding(String? name) => super.noSuchMethod(
         Invocation.method(
@@ -1509,6 +1668,7 @@ class MockHttpHeaders extends _i1.Mock implements _i2.HttpHeaders {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clear() => super.noSuchMethod(
         Invocation.method(
