@@ -1,4 +1,4 @@
-// Mocks generated by Mockito 5.4.0 from annotations
+// Mocks generated by Mockito 5.4.4 from annotations
 // in lxd_service/test/test_lxd_service.dart.
 // Do not manually edit this file.
 
@@ -7,11 +7,14 @@ import 'dart:async' as _i2;
 import 'dart:io' as _i3;
 
 import 'package:mockito/mockito.dart' as _i1;
+import 'package:mockito/src/dummies.dart' as _i4;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
@@ -56,51 +59,64 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   int get readyState => (super.noSuchMethod(
         Invocation.getter(#readyState),
         returnValue: 0,
       ) as int);
+
   @override
   String get extensions => (super.noSuchMethod(
         Invocation.getter(#extensions),
-        returnValue: '',
+        returnValue: _i4.dummyValue<String>(
+          this,
+          Invocation.getter(#extensions),
+        ),
       ) as String);
+
   @override
   bool get isBroadcast => (super.noSuchMethod(
         Invocation.getter(#isBroadcast),
         returnValue: false,
       ) as bool);
+
   @override
   _i2.Future<int> get length => (super.noSuchMethod(
         Invocation.getter(#length),
         returnValue: _i2.Future<int>.value(0),
       ) as _i2.Future<int>);
+
   @override
   _i2.Future<bool> get isEmpty => (super.noSuchMethod(
         Invocation.getter(#isEmpty),
         returnValue: _i2.Future<bool>.value(false),
       ) as _i2.Future<bool>);
+
   @override
   _i2.Future<dynamic> get first => (super.noSuchMethod(
         Invocation.getter(#first),
         returnValue: _i2.Future<dynamic>.value(),
       ) as _i2.Future<dynamic>);
+
   @override
   _i2.Future<dynamic> get last => (super.noSuchMethod(
         Invocation.getter(#last),
         returnValue: _i2.Future<dynamic>.value(),
       ) as _i2.Future<dynamic>);
+
   @override
   _i2.Future<dynamic> get single => (super.noSuchMethod(
         Invocation.getter(#single),
         returnValue: _i2.Future<dynamic>.value(),
       ) as _i2.Future<dynamic>);
+
   @override
   _i2.Future<dynamic> get done => (super.noSuchMethod(
         Invocation.getter(#done),
         returnValue: _i2.Future<dynamic>.value(),
       ) as _i2.Future<dynamic>);
+
   @override
   _i2.Future<dynamic> close([
     int? code,
@@ -116,6 +132,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Future<dynamic>.value(),
       ) as _i2.Future<dynamic>);
+
   @override
   void add(dynamic data) => super.noSuchMethod(
         Invocation.method(
@@ -124,6 +141,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i2.Future<dynamic> addStream(_i2.Stream<dynamic>? stream) =>
       (super.noSuchMethod(
@@ -133,6 +151,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Future<dynamic>.value(),
       ) as _i2.Future<dynamic>);
+
   @override
   void addUtf8Text(List<int>? bytes) => super.noSuchMethod(
         Invocation.method(
@@ -141,6 +160,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i2.Stream<dynamic> asBroadcastStream({
     void Function(_i2.StreamSubscription<dynamic>)? onListen,
@@ -157,6 +177,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Stream<dynamic>.empty(),
       ) as _i2.Stream<dynamic>);
+
   @override
   _i2.StreamSubscription<dynamic> listen(
     void Function(dynamic)? onData, {
@@ -187,6 +208,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
           ),
         ),
       ) as _i2.StreamSubscription<dynamic>);
+
   @override
   _i2.Stream<dynamic> where(bool Function(dynamic)? test) =>
       (super.noSuchMethod(
@@ -196,6 +218,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Stream<dynamic>.empty(),
       ) as _i2.Stream<dynamic>);
+
   @override
   _i2.Stream<S> map<S>(S Function(dynamic)? convert) => (super.noSuchMethod(
         Invocation.method(
@@ -204,6 +227,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Stream<S>.empty(),
       ) as _i2.Stream<S>);
+
   @override
   _i2.Stream<E> asyncMap<E>(_i2.FutureOr<E> Function(dynamic)? convert) =>
       (super.noSuchMethod(
@@ -213,6 +237,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Stream<E>.empty(),
       ) as _i2.Stream<E>);
+
   @override
   _i2.Stream<E> asyncExpand<E>(_i2.Stream<E>? Function(dynamic)? convert) =>
       (super.noSuchMethod(
@@ -222,6 +247,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Stream<E>.empty(),
       ) as _i2.Stream<E>);
+
   @override
   _i2.Stream<dynamic> handleError(
     Function? onError, {
@@ -235,6 +261,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Stream<dynamic>.empty(),
       ) as _i2.Stream<dynamic>);
+
   @override
   _i2.Stream<S> expand<S>(Iterable<S> Function(dynamic)? convert) =>
       (super.noSuchMethod(
@@ -244,6 +271,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Stream<S>.empty(),
       ) as _i2.Stream<S>);
+
   @override
   _i2.Future<dynamic> pipe(_i2.StreamConsumer<dynamic>? streamConsumer) =>
       (super.noSuchMethod(
@@ -253,6 +281,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Future<dynamic>.value(),
       ) as _i2.Future<dynamic>);
+
   @override
   _i2.Stream<S> transform<S>(
           _i2.StreamTransformer<dynamic, S>? streamTransformer) =>
@@ -263,6 +292,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Stream<S>.empty(),
       ) as _i2.Stream<S>);
+
   @override
   _i2.Future<dynamic> reduce(
           dynamic Function(
@@ -276,6 +306,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Future<dynamic>.value(),
       ) as _i2.Future<dynamic>);
+
   @override
   _i2.Future<S> fold<S>(
     S? initialValue,
@@ -292,25 +323,46 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
             combine,
           ],
         ),
-        returnValue: _FakeFuture_1<S>(
-          this,
-          Invocation.method(
-            #fold,
-            [
-              initialValue,
-              combine,
-            ],
-          ),
-        ),
+        returnValue: _i4.ifNotNull(
+              _i4.dummyValueOrNull<S>(
+                this,
+                Invocation.method(
+                  #fold,
+                  [
+                    initialValue,
+                    combine,
+                  ],
+                ),
+              ),
+              (S v) => _i2.Future<S>.value(v),
+            ) ??
+            _FakeFuture_1<S>(
+              this,
+              Invocation.method(
+                #fold,
+                [
+                  initialValue,
+                  combine,
+                ],
+              ),
+            ),
       ) as _i2.Future<S>);
+
   @override
   _i2.Future<String> join([String? separator = r'']) => (super.noSuchMethod(
         Invocation.method(
           #join,
           [separator],
         ),
-        returnValue: _i2.Future<String>.value(''),
+        returnValue: _i2.Future<String>.value(_i4.dummyValue<String>(
+          this,
+          Invocation.method(
+            #join,
+            [separator],
+          ),
+        )),
       ) as _i2.Future<String>);
+
   @override
   _i2.Future<bool> contains(Object? needle) => (super.noSuchMethod(
         Invocation.method(
@@ -319,6 +371,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Future<bool>.value(false),
       ) as _i2.Future<bool>);
+
   @override
   _i2.Future<void> forEach(void Function(dynamic)? action) =>
       (super.noSuchMethod(
@@ -329,6 +382,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         returnValue: _i2.Future<void>.value(),
         returnValueForMissingStub: _i2.Future<void>.value(),
       ) as _i2.Future<void>);
+
   @override
   _i2.Future<bool> every(bool Function(dynamic)? test) => (super.noSuchMethod(
         Invocation.method(
@@ -337,6 +391,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Future<bool>.value(false),
       ) as _i2.Future<bool>);
+
   @override
   _i2.Future<bool> any(bool Function(dynamic)? test) => (super.noSuchMethod(
         Invocation.method(
@@ -345,6 +400,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Future<bool>.value(false),
       ) as _i2.Future<bool>);
+
   @override
   _i2.Stream<R> cast<R>() => (super.noSuchMethod(
         Invocation.method(
@@ -353,6 +409,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Stream<R>.empty(),
       ) as _i2.Stream<R>);
+
   @override
   _i2.Future<List<dynamic>> toList() => (super.noSuchMethod(
         Invocation.method(
@@ -361,6 +418,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Future<List<dynamic>>.value(<dynamic>[]),
       ) as _i2.Future<List<dynamic>>);
+
   @override
   _i2.Future<Set<dynamic>> toSet() => (super.noSuchMethod(
         Invocation.method(
@@ -369,20 +427,32 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Future<Set<dynamic>>.value(<dynamic>{}),
       ) as _i2.Future<Set<dynamic>>);
+
   @override
   _i2.Future<E> drain<E>([E? futureValue]) => (super.noSuchMethod(
         Invocation.method(
           #drain,
           [futureValue],
         ),
-        returnValue: _FakeFuture_1<E>(
-          this,
-          Invocation.method(
-            #drain,
-            [futureValue],
-          ),
-        ),
+        returnValue: _i4.ifNotNull(
+              _i4.dummyValueOrNull<E>(
+                this,
+                Invocation.method(
+                  #drain,
+                  [futureValue],
+                ),
+              ),
+              (E v) => _i2.Future<E>.value(v),
+            ) ??
+            _FakeFuture_1<E>(
+              this,
+              Invocation.method(
+                #drain,
+                [futureValue],
+              ),
+            ),
       ) as _i2.Future<E>);
+
   @override
   _i2.Stream<dynamic> take(int? count) => (super.noSuchMethod(
         Invocation.method(
@@ -391,6 +461,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Stream<dynamic>.empty(),
       ) as _i2.Stream<dynamic>);
+
   @override
   _i2.Stream<dynamic> takeWhile(bool Function(dynamic)? test) =>
       (super.noSuchMethod(
@@ -400,6 +471,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Stream<dynamic>.empty(),
       ) as _i2.Stream<dynamic>);
+
   @override
   _i2.Stream<dynamic> skip(int? count) => (super.noSuchMethod(
         Invocation.method(
@@ -408,6 +480,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Stream<dynamic>.empty(),
       ) as _i2.Stream<dynamic>);
+
   @override
   _i2.Stream<dynamic> skipWhile(bool Function(dynamic)? test) =>
       (super.noSuchMethod(
@@ -417,6 +490,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Stream<dynamic>.empty(),
       ) as _i2.Stream<dynamic>);
+
   @override
   _i2.Stream<dynamic> distinct(
           [bool Function(
@@ -430,6 +504,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Stream<dynamic>.empty(),
       ) as _i2.Stream<dynamic>);
+
   @override
   _i2.Future<dynamic> firstWhere(
     bool Function(dynamic)? test, {
@@ -443,6 +518,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Future<dynamic>.value(),
       ) as _i2.Future<dynamic>);
+
   @override
   _i2.Future<dynamic> lastWhere(
     bool Function(dynamic)? test, {
@@ -456,6 +532,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Future<dynamic>.value(),
       ) as _i2.Future<dynamic>);
+
   @override
   _i2.Future<dynamic> singleWhere(
     bool Function(dynamic)? test, {
@@ -469,6 +546,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Future<dynamic>.value(),
       ) as _i2.Future<dynamic>);
+
   @override
   _i2.Future<dynamic> elementAt(int? index) => (super.noSuchMethod(
         Invocation.method(
@@ -477,6 +555,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Future<dynamic>.value(),
       ) as _i2.Future<dynamic>);
+
   @override
   _i2.Stream<dynamic> timeout(
     Duration? timeLimit, {
@@ -490,6 +569,7 @@ class MockWebSocket extends _i1.Mock implements _i3.WebSocket {
         ),
         returnValue: _i2.Stream<dynamic>.empty(),
       ) as _i2.Stream<dynamic>);
+
   @override
   void addError(
     Object? error, [
