@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:movable_tabs/movable_tabs.dart';
 import 'package:movable_tabs/src/tab_x.dart';
-import 'package:yaru_icons/yaru_icons.dart';
+import 'package:yaru/icons.dart';
 
 void main() {
   Widget buildTabs(
