@@ -1,19 +1,22 @@
-// Mocks generated by Mockito 5.4.0 from annotations
+// Mocks generated by Mockito 5.4.4 from annotations
 // in os_logo/test/os_logo_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i3;
-import 'dart:typed_data' as _i5;
-import 'dart:ui' as _i2;
+import 'dart:async' as _i2;
+import 'dart:typed_data' as _i4;
+import 'dart:ui' as _i5;
 
-import 'package:flutter/src/services/asset_bundle.dart' as _i4;
+import 'package:flutter/src/services/asset_bundle.dart' as _i3;
 import 'package:mockito/mockito.dart' as _i1;
+import 'package:mockito/src/dummies.dart' as _i6;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
@@ -21,19 +24,8 @@ import 'package:mockito/mockito.dart' as _i1;
 // ignore_for_file: camel_case_types
 // ignore_for_file: subtype_of_sealed_class
 
-class _FakeImmutableBuffer_0 extends _i1.SmartFake
-    implements _i2.ImmutableBuffer {
-  _FakeImmutableBuffer_0(
-    Object parent,
-    Invocation parentInvocation,
-  ) : super(
-          parent,
-          parentInvocation,
-        );
-}
-
-class _FakeFuture_1<T> extends _i1.SmartFake implements _i3.Future<T> {
-  _FakeFuture_1(
+class _FakeFuture_0<T1> extends _i1.SmartFake implements _i2.Future<T1> {
+  _FakeFuture_0(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -45,37 +37,39 @@ class _FakeFuture_1<T> extends _i1.SmartFake implements _i3.Future<T> {
 /// A class which mocks [AssetBundle].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockAssetBundle extends _i1.Mock implements _i4.AssetBundle {
+class MockAssetBundle extends _i1.Mock implements _i3.AssetBundle {
   MockAssetBundle() {
     _i1.throwOnMissingStub(this);
   }
 
   @override
-  _i3.Future<_i5.ByteData> load(String? key) => (super.noSuchMethod(
+  _i2.Future<_i4.ByteData> load(String? key) => (super.noSuchMethod(
         Invocation.method(
           #load,
           [key],
         ),
-        returnValue: _i3.Future<_i5.ByteData>.value(_i5.ByteData(0)),
-      ) as _i3.Future<_i5.ByteData>);
+        returnValue: _i2.Future<_i4.ByteData>.value(_i4.ByteData(0)),
+      ) as _i2.Future<_i4.ByteData>);
+
   @override
-  _i3.Future<_i2.ImmutableBuffer> loadBuffer(String? key) =>
+  _i2.Future<_i5.ImmutableBuffer> loadBuffer(String? key) =>
       (super.noSuchMethod(
         Invocation.method(
           #loadBuffer,
           [key],
         ),
-        returnValue:
-            _i3.Future<_i2.ImmutableBuffer>.value(_FakeImmutableBuffer_0(
+        returnValue: _i2.Future<_i5.ImmutableBuffer>.value(
+            _i6.dummyValue<_i5.ImmutableBuffer>(
           this,
           Invocation.method(
             #loadBuffer,
             [key],
           ),
         )),
-      ) as _i3.Future<_i2.ImmutableBuffer>);
+      ) as _i2.Future<_i5.ImmutableBuffer>);
+
   @override
-  _i3.Future<String> loadString(
+  _i2.Future<String> loadString(
     String? key, {
     bool? cache = true,
   }) =>
@@ -85,12 +79,20 @@ class MockAssetBundle extends _i1.Mock implements _i4.AssetBundle {
           [key],
           {#cache: cache},
         ),
-        returnValue: _i3.Future<String>.value(''),
-      ) as _i3.Future<String>);
+        returnValue: _i2.Future<String>.value(_i6.dummyValue<String>(
+          this,
+          Invocation.method(
+            #loadString,
+            [key],
+            {#cache: cache},
+          ),
+        )),
+      ) as _i2.Future<String>);
+
   @override
-  _i3.Future<T> loadStructuredData<T>(
+  _i2.Future<T> loadStructuredData<T>(
     String? key,
-    _i3.Future<T> Function(String)? parser,
+    _i2.Future<T> Function(String)? parser,
   ) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -100,21 +102,35 @@ class MockAssetBundle extends _i1.Mock implements _i4.AssetBundle {
             parser,
           ],
         ),
-        returnValue: _FakeFuture_1<T>(
-          this,
-          Invocation.method(
-            #loadStructuredData,
-            [
-              key,
-              parser,
-            ],
-          ),
-        ),
-      ) as _i3.Future<T>);
+        returnValue: _i6.ifNotNull(
+              _i6.dummyValueOrNull<T>(
+                this,
+                Invocation.method(
+                  #loadStructuredData,
+                  [
+                    key,
+                    parser,
+                  ],
+                ),
+              ),
+              (T v) => _i2.Future<T>.value(v),
+            ) ??
+            _FakeFuture_0<T>(
+              this,
+              Invocation.method(
+                #loadStructuredData,
+                [
+                  key,
+                  parser,
+                ],
+              ),
+            ),
+      ) as _i2.Future<T>);
+
   @override
-  _i3.Future<T> loadStructuredBinaryData<T>(
+  _i2.Future<T> loadStructuredBinaryData<T>(
     String? key,
-    _i3.FutureOr<T> Function(_i5.ByteData)? parser,
+    _i2.FutureOr<T> Function(_i4.ByteData)? parser,
   ) =>
       (super.noSuchMethod(
         Invocation.method(
@@ -124,17 +140,31 @@ class MockAssetBundle extends _i1.Mock implements _i4.AssetBundle {
             parser,
           ],
         ),
-        returnValue: _FakeFuture_1<T>(
-          this,
-          Invocation.method(
-            #loadStructuredBinaryData,
-            [
-              key,
-              parser,
-            ],
-          ),
-        ),
-      ) as _i3.Future<T>);
+        returnValue: _i6.ifNotNull(
+              _i6.dummyValueOrNull<T>(
+                this,
+                Invocation.method(
+                  #loadStructuredBinaryData,
+                  [
+                    key,
+                    parser,
+                  ],
+                ),
+              ),
+              (T v) => _i2.Future<T>.value(v),
+            ) ??
+            _FakeFuture_0<T>(
+              this,
+              Invocation.method(
+                #loadStructuredBinaryData,
+                [
+                  key,
+                  parser,
+                ],
+              ),
+            ),
+      ) as _i2.Future<T>);
+
   @override
   void evict(String? key) => super.noSuchMethod(
         Invocation.method(
@@ -143,6 +173,7 @@ class MockAssetBundle extends _i1.Mock implements _i4.AssetBundle {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   void clear() => super.noSuchMethod(
         Invocation.method(
