@@ -1,4 +1,4 @@
-// Mocks generated by Mockito 5.4.0 from annotations
+// Mocks generated by Mockito 5.4.4 from annotations
 // in workshops/test/instance_info_model_test.dart.
 // Do not manually edit this file.
 
@@ -7,15 +7,18 @@ import 'dart:async' as _i6;
 import 'dart:io' as _i3;
 
 import 'package:lxd/lxd.dart' as _i2;
-import 'package:lxd_service/src/features.dart' as _i7;
+import 'package:lxd_service/src/features.dart' as _i8;
 import 'package:lxd_service/src/service.dart' as _i5;
 import 'package:lxd_service/src/terminal.dart' as _i4;
 import 'package:mockito/mockito.dart' as _i1;
+import 'package:mockito/src/dummies.dart' as _i7;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
 // ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
@@ -214,6 +217,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           Invocation.getter(#url),
         ),
       ) as Uri);
+
   @override
   set userAgent(String? value) => super.noSuchMethod(
         Invocation.setter(
@@ -222,26 +226,31 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i6.Stream<List<_i2.LxdInstanceId>> get instanceStream => (super.noSuchMethod(
         Invocation.getter(#instanceStream),
         returnValue: _i6.Stream<List<_i2.LxdInstanceId>>.empty(),
       ) as _i6.Stream<List<_i2.LxdInstanceId>>);
+
   @override
   _i6.Stream<_i2.LxdInstanceId> get instanceAdded => (super.noSuchMethod(
         Invocation.getter(#instanceAdded),
         returnValue: _i6.Stream<_i2.LxdInstanceId>.empty(),
       ) as _i6.Stream<_i2.LxdInstanceId>);
+
   @override
   _i6.Stream<_i2.LxdInstanceId> get instanceRemoved => (super.noSuchMethod(
         Invocation.getter(#instanceRemoved),
         returnValue: _i6.Stream<_i2.LxdInstanceId>.empty(),
       ) as _i6.Stream<_i2.LxdInstanceId>);
+
   @override
   _i6.Stream<_i2.LxdInstanceId> get instanceUpdated => (super.noSuchMethod(
         Invocation.getter(#instanceUpdated),
         returnValue: _i6.Stream<_i2.LxdInstanceId>.empty(),
       ) as _i6.Stream<_i2.LxdInstanceId>);
+
   @override
   _i6.Future<void> init() => (super.noSuchMethod(
         Invocation.method(
@@ -251,6 +260,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         returnValue: _i6.Future<void>.value(),
         returnValueForMissingStub: _i6.Future<void>.value(),
       ) as _i6.Future<void>);
+
   @override
   _i6.Future<Map<String, List<String>>> getOperations() => (super.noSuchMethod(
         Invocation.method(
@@ -260,6 +270,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         returnValue: _i6.Future<Map<String, List<String>>>.value(
             <String, List<String>>{}),
       ) as _i6.Future<Map<String, List<String>>>);
+
   @override
   _i6.Future<_i2.LxdServer> getServerInfo({String? target}) =>
       (super.noSuchMethod(
@@ -277,6 +288,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdServer>);
+
   @override
   _i6.Future<_i2.LxdOperation> getOperation(String? id) => (super.noSuchMethod(
         Invocation.method(
@@ -291,6 +303,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdOperation>);
+
   @override
   _i6.Future<_i3.WebSocket> getOperationWebSocket(
     String? id,
@@ -315,6 +328,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i3.WebSocket>);
+
   @override
   _i6.Future<_i2.LxdOperation> waitOperation(
     String? id, {
@@ -335,6 +349,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdOperation>);
+
   @override
   _i6.Future<void> cancelOperation(String? id) => (super.noSuchMethod(
         Invocation.method(
@@ -344,6 +359,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         returnValue: _i6.Future<void>.value(),
         returnValueForMissingStub: _i6.Future<void>.value(),
       ) as _i6.Future<void>);
+
   @override
   _i6.Future<_i2.LxdResources> getResources() => (super.noSuchMethod(
         Invocation.method(
@@ -358,6 +374,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdResources>);
+
   @override
   _i6.Future<List<String>> getCertificates() => (super.noSuchMethod(
         Invocation.method(
@@ -366,6 +383,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         ),
         returnValue: _i6.Future<List<String>>.value(<String>[]),
       ) as _i6.Future<List<String>>);
+
   @override
   _i6.Future<_i2.LxdCertificate> getCertificate(String? fingerprint) =>
       (super.noSuchMethod(
@@ -381,6 +399,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdCertificate>);
+
   @override
   _i6.Stream<_i2.LxdEvent> getEvents({
     String? project,
@@ -397,6 +416,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         ),
         returnValue: _i6.Stream<_i2.LxdEvent>.empty(),
       ) as _i6.Stream<_i2.LxdEvent>);
+
   @override
   _i6.Stream<_i2.LxdEvent> getAllEvents(
           {Set<_i2.LxdEventType>? types = const {}}) =>
@@ -408,6 +428,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         ),
         returnValue: _i6.Stream<_i2.LxdEvent>.empty(),
       ) as _i6.Stream<_i2.LxdEvent>);
+
   @override
   _i6.Future<List<String>> getImages({
     String? project,
@@ -424,6 +445,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         ),
         returnValue: _i6.Future<List<String>>.value(<String>[]),
       ) as _i6.Future<List<String>>);
+
   @override
   _i6.Future<_i2.LxdImage> getImage(
     String? fingerprint, {
@@ -444,6 +466,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdImage>);
+
   @override
   _i6.Future<List<_i2.LxdInstanceId>> getInstances({
     String? project,
@@ -461,6 +484,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         returnValue:
             _i6.Future<List<_i2.LxdInstanceId>>.value(<_i2.LxdInstanceId>[]),
       ) as _i6.Future<List<_i2.LxdInstanceId>>);
+
   @override
   _i6.Future<List<_i2.LxdInstanceId>> getAllInstances({String? filter}) =>
       (super.noSuchMethod(
@@ -472,6 +496,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         returnValue:
             _i6.Future<List<_i2.LxdInstanceId>>.value(<_i2.LxdInstanceId>[]),
       ) as _i6.Future<List<_i2.LxdInstanceId>>);
+
   @override
   _i6.Future<_i2.LxdInstance> getInstance(_i2.LxdInstanceId? id) =>
       (super.noSuchMethod(
@@ -487,6 +512,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdInstance>);
+
   @override
   _i6.Future<_i2.LxdInstanceState> getInstanceState(_i2.LxdInstanceId? id) =>
       (super.noSuchMethod(
@@ -503,6 +529,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdInstanceState>);
+
   @override
   _i6.Future<_i2.LxdOperation> createInstance({
     String? project,
@@ -565,6 +592,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdOperation>);
+
   @override
   _i6.Future<_i2.LxdOperation> startInstance(
     _i2.LxdInstanceId? id, {
@@ -585,6 +613,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdOperation>);
+
   @override
   _i6.Future<_i2.LxdOperation> execInstance(
     _i2.LxdInstanceId? id, {
@@ -636,6 +665,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdOperation>);
+
   @override
   _i6.Future<_i2.LxdOperation> updateInstance(_i2.LxdInstance? instance) =>
       (super.noSuchMethod(
@@ -651,6 +681,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdOperation>);
+
   @override
   _i6.Future<_i2.LxdOperation> stopInstance(
     _i2.LxdInstanceId? id, {
@@ -678,6 +709,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdOperation>);
+
   @override
   _i6.Future<_i2.LxdOperation> restartInstance(
     _i2.LxdInstanceId? id, {
@@ -705,6 +737,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdOperation>);
+
   @override
   _i6.Future<_i2.LxdOperation> deleteInstance(_i2.LxdInstanceId? id) =>
       (super.noSuchMethod(
@@ -720,6 +753,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdOperation>);
+
   @override
   _i6.Future<String> pullFile(
     _i2.LxdInstanceId? id, {
@@ -731,8 +765,16 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           [id],
           {#path: path},
         ),
-        returnValue: _i6.Future<String>.value(''),
+        returnValue: _i6.Future<String>.value(_i7.dummyValue<String>(
+          this,
+          Invocation.method(
+            #pullFile,
+            [id],
+            {#path: path},
+          ),
+        )),
       ) as _i6.Future<String>);
+
   @override
   _i6.Future<void> deleteFile(
     _i2.LxdInstanceId? id, {
@@ -747,6 +789,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         returnValue: _i6.Future<void>.value(),
         returnValueForMissingStub: _i6.Future<void>.value(),
       ) as _i6.Future<void>);
+
   @override
   _i6.Future<void> pushFile(
     _i2.LxdInstanceId? id, {
@@ -775,6 +818,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         returnValue: _i6.Future<void>.value(),
         returnValueForMissingStub: _i6.Future<void>.value(),
       ) as _i6.Future<void>);
+
   @override
   _i6.Future<List<String>> getNetworks() => (super.noSuchMethod(
         Invocation.method(
@@ -783,6 +827,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         ),
         returnValue: _i6.Future<List<String>>.value(<String>[]),
       ) as _i6.Future<List<String>>);
+
   @override
   _i6.Future<_i2.LxdNetwork> getNetwork(String? name) => (super.noSuchMethod(
         Invocation.method(
@@ -797,6 +842,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdNetwork>);
+
   @override
   _i6.Future<List<_i2.LxdNetworkLease>> getNetworkLeases(String? name) =>
       (super.noSuchMethod(
@@ -807,6 +853,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         returnValue: _i6.Future<List<_i2.LxdNetworkLease>>.value(
             <_i2.LxdNetworkLease>[]),
       ) as _i6.Future<List<_i2.LxdNetworkLease>>);
+
   @override
   _i6.Future<_i2.LxdNetworkState> getNetworkState(String? name) =>
       (super.noSuchMethod(
@@ -823,6 +870,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdNetworkState>);
+
   @override
   _i6.Future<List<String>> getNetworkAcls({String? project}) =>
       (super.noSuchMethod(
@@ -833,6 +881,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         ),
         returnValue: _i6.Future<List<String>>.value(<String>[]),
       ) as _i6.Future<List<String>>);
+
   @override
   _i6.Future<_i2.LxdNetworkAcl> getNetworkAcl(
     String? name, {
@@ -853,6 +902,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdNetworkAcl>);
+
   @override
   _i6.Future<List<String>> getProfiles({String? project}) =>
       (super.noSuchMethod(
@@ -863,6 +913,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         ),
         returnValue: _i6.Future<List<String>>.value(<String>[]),
       ) as _i6.Future<List<String>>);
+
   @override
   _i6.Future<_i2.LxdProfile> getProfile(
     String? name, {
@@ -883,6 +934,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdProfile>);
+
   @override
   _i6.Future<List<String>> getProjects() => (super.noSuchMethod(
         Invocation.method(
@@ -891,6 +943,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         ),
         returnValue: _i6.Future<List<String>>.value(<String>[]),
       ) as _i6.Future<List<String>>);
+
   @override
   _i6.Future<_i2.LxdProject> getProject(String? name) => (super.noSuchMethod(
         Invocation.method(
@@ -905,6 +958,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdProject>);
+
   @override
   _i6.Future<_i2.LxdProjectState> getProjectState(String? name) =>
       (super.noSuchMethod(
@@ -921,6 +975,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdProjectState>);
+
   @override
   _i6.Future<void> updateProject(_i2.LxdProject? project) =>
       (super.noSuchMethod(
@@ -931,6 +986,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         returnValue: _i6.Future<void>.value(),
         returnValueForMissingStub: _i6.Future<void>.value(),
       ) as _i6.Future<void>);
+
   @override
   _i6.Future<List<String>> getStoragePools({String? project}) =>
       (super.noSuchMethod(
@@ -941,6 +997,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         ),
         returnValue: _i6.Future<List<String>>.value(<String>[]),
       ) as _i6.Future<List<String>>);
+
   @override
   _i6.Future<_i2.LxdStoragePool> getStoragePool(
     String? name, {
@@ -962,6 +1019,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdStoragePool>);
+
   @override
   void close() => super.noSuchMethod(
         Invocation.method(
@@ -970,10 +1028,11 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         ),
         returnValueForMissingStub: null,
       );
+
   @override
   _i6.Future<_i2.LxdOperation?> initFeature(
     _i2.LxdInstanceId? id,
-    _i7.LxdFeatureProvider? feature,
+    _i8.LxdFeatureProvider? feature,
     _i2.LxdImage? image,
   ) =>
       (super.noSuchMethod(
@@ -987,6 +1046,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         ),
         returnValue: _i6.Future<_i2.LxdOperation?>.value(),
       ) as _i6.Future<_i2.LxdOperation?>);
+
   @override
   _i6.Future<_i2.LxdImage> configureImage(
     _i2.LxdInstanceId? id,
@@ -1011,10 +1071,11 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdImage>);
+
   @override
   _i6.Future<void> configureFeature(
     _i2.LxdInstanceId? id,
-    _i7.LxdFeatureProvider? feature,
+    _i8.LxdFeatureProvider? feature,
     _i2.LxdImage? image,
   ) =>
       (super.noSuchMethod(
@@ -1029,10 +1090,11 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         returnValue: _i6.Future<void>.value(),
         returnValueForMissingStub: _i6.Future<void>.value(),
       ) as _i6.Future<void>);
+
   @override
   _i6.Future<_i2.LxdOperation> stageFeatures(
     _i2.LxdInstanceId? id,
-    List<_i7.LxdFeatureProvider>? features,
+    List<_i8.LxdFeatureProvider>? features,
     _i2.LxdImage? image,
   ) =>
       (super.noSuchMethod(
@@ -1056,6 +1118,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdOperation>);
+
   @override
   _i6.Future<void> dispose() => (super.noSuchMethod(
         Invocation.method(
@@ -1065,6 +1128,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         returnValue: _i6.Future<void>.value(),
         returnValueForMissingStub: _i6.Future<void>.value(),
       ) as _i6.Future<void>);
+
   @override
   _i6.Future<_i2.LxdOperation> updateInstanceConfig(
     _i2.LxdInstanceId? id,
@@ -1089,6 +1153,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i2.LxdOperation>);
+
   @override
   _i6.Future<void> updateProjectConfig(
     String? name,
@@ -1105,6 +1170,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         returnValue: _i6.Future<void>.value(),
         returnValueForMissingStub: _i6.Future<void>.value(),
       ) as _i6.Future<void>);
+
   @override
   _i6.Future<_i4.LxdTerminal> execTerminal(_i2.LxdInstanceId? id) =>
       (super.noSuchMethod(
@@ -1120,6 +1186,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
           ),
         )),
       ) as _i6.Future<_i4.LxdTerminal>);
+
   @override
   _i6.Future<bool> waitVmAgent(
     _i2.LxdInstanceId? id, {
@@ -1137,6 +1204,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         ),
         returnValue: _i6.Future<bool>.value(false),
       ) as _i6.Future<bool>);
+
   @override
   _i6.Stream<_i2.LxdOperation> watchInstance(_i2.LxdInstanceId? id) =>
       (super.noSuchMethod(
@@ -1146,6 +1214,7 @@ class MockLxdService extends _i1.Mock implements _i5.LxdService {
         ),
         returnValue: _i6.Stream<_i2.LxdOperation>.empty(),
       ) as _i6.Stream<_i2.LxdOperation>);
+
   @override
   _i6.Stream<_i2.LxdOperation> watchOperation(String? id) =>
       (super.noSuchMethod(
